Struct tslib::i2c::I2c

pub struct I2c<S, R> where
    S: Any + I2C,
    R: IsRemapped,  { /* fields omitted */ }

Methods

impl<S, R> I2c<S, R> where
S: Any + I2C,
R: IsRemapped,

pub fn new<F>(
i2c: S,
ports: I2cBusPorts<S, R>,
freq: F,
clocks: Clocks,
duty: I2cDutyCycle
) -> Self where
F: Into<Hertz>,

Initializes the given I2c module

Arguments

i2c: The I2C module.

freq: The I2c bus frequency, typically either 100 kHz (slow mode) or 400 kHz (fast mode)

duty: The duty cycle of the I2c module.

pub fn listen(&mut self)

Enable all I2C module interrupts.

pub fn start(&mut self) -> Result<(), Error>

Try to generate a start signal. Complete using the start_complete function.

pub fn stop(&mut self) -> Result<(), Error>

Give up control over the I2C bus.

pub fn start_complete(&self) -> Result<(), Error>

Returns Ok(()) once the I2C module has finished transmitting the start signal, otherwise returns Err(nb::Error::WouldBlock).

Remarks

Use block!(i2c.start_complete()).unwrap() to continue polling until the start signal is generated.

pub fn addr_complete(&self) -> Result<(), Error>

Returns Ok(()) if address has finished sending (master) or if matching address has been received (slave). Returns Err(nb::Error::WouldBlock) otherwise.

Remarks

Use block!(i2c.addr_complete()).unwrap() to continue polling until event happens.

pub fn transmit_empty(&self) -> Result<(), Error>

Returns Ok(()) if the transmit buffer is empty. Returns Err(nb::Error::WouldBlock) otherwise.

Remarks

Use block!(i2c.transmit_empty()).unwrap() to continue polling until event happens.

pub fn receive_not_empty(&self) -> Result<(), Error>

Returns Ok(()) if the receive buffer is not empty. Returns Err(nb::Error::WouldBlock) otherwise.

Remarks

Use block!(i2c.receive_not_empty()).unwrap() to continue polling until event happens.

pub unsafe fn write_unchecked(&mut self, dat: u8) -> Result<(), Error>

Writes to the I2C data register. This function should most likely not be used directly.

pub unsafe fn read_unchecked(&mut self) -> Result<u8, Error>

Reads from the I2C data register. This should not be used directly.

impl I2c<I2C1, NotRemapped>

pub fn ports_normal<M>(
_pb6: GpioPin<GPIOB, Pin6, M, PinCnf3>,
_pb7: GpioPin<GPIOB, Pin7, M, PinCnf3>,
_afio_i2c: AfioPeripheral<I2C1, NotRemapped>
) -> I2cBusPorts<I2C1, NotRemapped> where
M: PinOutput + PinMode,

impl I2c<I2C1, Remapped>

pub fn ports_remapped<M>(
_pb8: GpioPin<GPIOB, Pin8, M, PinCnf3>,
_pb9: GpioPin<GPIOB, Pin9, M, PinCnf3>,
_afio_i2c: AfioPeripheral<I2C1, Remapped>
) -> I2cBusPorts<I2C1, Remapped> where
M: PinOutput + PinMode,